Simplify VizHub Theme
